+++ test/pow/extension/ecto/schema_test.exs
@@ -59,6 +59,8 @@ defmodule Pow.Extension.Ecto.SchemaTest do
     use Pow.Extension.Ecto.Schema,
       extensions: [Pow.Extension.Ecto.SchemaTest.ExtensionMock]
 
+    @ecto_derive_inspect_for_redacted_fields false
+
     schema "users" do
       pow_user_fields()
 
@@ -80,6 +82,8 @@ defmodule Pow.Extension.Ecto.SchemaTest do
           user_id_field: :username
         use Pow.Extension.Ecto.Schema,
           extensions: [Pow.Extension.Ecto.SchemaTest.ExtensionMock]
+
+        @ecto_derive_inspect_for_redacted_fields false
 
         schema "users" do
           pow_user_fields()