+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v1.0.25 (TBA)
 
+### Enhancements
+
+* [`Pow.Ecto.Schema.Fields`] The `:password_hash`, `:current_password`, and `:password` fields now have `redact: true` option set
+
 ### Bug fixes
 
 * [`Pow.Operations`] `Pow.Operations.fetch_primary_key_values/2` now ensures that module exists and is loaded before deriving primary keys